[pycti/tests] Migrated work related functions from test utils to OpenCTIApiWork
