+++ pycti/api/opencti_api_work.py
@@ -1,4 +1,6 @@
 import logging
+import time
+from typing import Dict, List
 
 
 class OpenCTIApiWork:
@@ -72,3 +74,108 @@ class OpenCTIApiWork:
             query, {"connectorId": connector_id, "friendlyName": friendly_name}
         )
         return work["data"]["workAdd"]["id"]
+
+    def delete_work(self, work_id: str):
+        query = """
+        mutation ConnectorWorksMutation($workId: ID!) {
+            workEdit(id: $workId) {
+                delete
+            }
+        }"""
+        work = self.api.query(
+            query,
+            {"workId": work_id},
+        )
+        return work["data"]
+
+    def wait_for_work_to_finish(self, connector_id: str, work_id: str):
+        status = ""
+        cnt = 0
+        while status != "complete":
+            states = self.get_connector_works(
+                connector_id=connector_id, work_id=work_id
+            )
+            if len(states) > 0:
+                status = states[0]["status"]
+
+                if states[0]["errors"]:
+                    self.api.log(
+                        "error", f"Unexpected connector error {states[0]['errors']}"
+                    )
+                    return ""
+
+            time.sleep(1)
+            cnt += 1
+
+    def get_connector_works(self, connector_id: str, work_id: str = "") -> List[Dict]:
+        if connector_id == "" and work_id == "":
+            self.api.log("error", "Missing argument: connector_id or work_id")
+            return [{}]
+
+        query = """
+            query ConnectorWorksQuery(
+                  $count: Int
+                  $orderBy: WorksOrdering
+                  $orderMode: OrderingMode
+                  $filters: [WorksFiltering]
+                ) {
+                  works(
+                    first: $count
+                    orderBy: $orderBy
+                    orderMode: $orderMode
+                    filters: $filters
+                  ) {
+                    edges {
+                      node {
+                        id
+                        name
+                        user {
+                          name
+                        }
+                        timestamp
+                        status
+                        event_source_id
+                        received_time
+                        processed_time
+                        completed_time
+                        tracking {
+                          import_expected_number
+                          import_processed_number
+                        }
+                        messages {
+                          timestamp
+                          message
+                          sequence
+                          source
+                        }
+                        errors {
+                          timestamp
+                          message
+                          sequence
+                          source
+                        }
+                      }
+                    }
+                  }
+                }
+                """
+        result = self.api.query(
+            query,
+            {
+                "count": 50,
+                "filters": [
+                    {"key": "connector_id", "values": [connector_id]},
+                ],
+            },
+        )
+        result = result["data"]["works"]["edges"]
+        return_value = []
+        for node in result:
+            node = node["node"]
+            if work_id != "":
+                if node["id"] == work_id:
+                    return_value.append(node)
+            else:
+                return_value.append(node)
+
+        return sorted(return_value, key=lambda i: i["timestamp"])
